pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolio-api-test-pack"
version = "0.1.0"
description = "Realistic pytest and Allure API test pack scaffold"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "allure-pytest>=2.15.3",
  "jsonschema>=4.23.0",
  "PyYAML>=6.0.2",
  "pytest>=8.3.5",
  "requests>=2.32.3",
  "responses>=0.25.7",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["tests*"]

[tool.pytest.ini_options]
addopts = "-ra --strict-markers"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
